+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the "auto-myrient" project will be documented in this fil
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Retry logic for failed downloads
+
 ## [v0.1.0] - 2024-07-11
 
 ### Added